Set Boat possible allures
