+++ Assets/Scripts/Boat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fed059f66f9e8fe4fbef159dad447bfc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
