+++ Assets/Scripts/Boat/CameraRotaion.cs
@@ -9,9 +9,9 @@ public class CameraRotaion : MonoBehaviour
 
     private void Update()
     {
-        if (_rotationAxis != Vector2.zero)
+        if (_rotationAxis != Vector2.zero && Input.GetKey(KeyCode.Mouse0))
         {
-            _mainCameraPivot.localEulerAngles += new Vector3(0, _rotationAxis.x * _cameraRotationSpeed * Time.deltaTime, 0);
+            _mainCameraPivot.Rotate(new Vector3(0, _rotationAxis.x * _cameraRotationSpeed * Time.deltaTime, 0));
         }
     }
