+++ Assets/Scripts/GameManager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 78045d3cba96cd747bc8077fefb1b2db
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
